Also pass the caller option to loadPartialConfig
