+++ src/transform.js
@@ -1,6 +1,7 @@
 const babel = require("@babel/core");
 const promisify = require("util.promisify");
 const LoaderError = require("./Error");
+const injectCaller = require("./injectCaller");
 
 const transform = promisify(babel.transform);
 
@@ -29,43 +30,3 @@ module.exports = async function(source, options) {
 };
 
 module.exports.version = babel.version;
-
-function injectCaller(opts) {
-  if (!supportsCallerOption()) return opts;
-
-  return Object.assign({}, opts, {
-    caller: Object.assign(
-      {
-        name: "babel-loader",
-
-        // Webpack >= 2 supports ESM and dynamic import.
-        supportsStaticESM: true,
-        supportsDynamicImport: true,
-      },
-      opts.caller,
-    ),
-  });
-}
-
-// TODO: We can remove this eventually, I'm just adding it so that people have
-// a little time to migrate to the newer RCs of @babel/core without getting
-// hard-to-diagnose errors about unknown 'caller' options.
-let supportsCallerOptionFlag = undefined;
-function supportsCallerOption() {
-  if (supportsCallerOptionFlag === undefined) {
-    try {
-      // Rather than try to match the Babel version, we just see if it throws
-      // when passed a 'caller' flag, and use that to decide if it is supported.
-      babel.loadPartialConfig({
-        caller: undefined,
-        babelrc: false,
-        configFile: false,
-      });
-      supportsCallerOptionFlag = true;
-    } catch (err) {
-      supportsCallerOptionFlag = false;
-    }
-  }
-
-  return supportsCallerOptionFlag;
-}
